Move ResolvedMethod into compiler/compile
Compiler extensions have to define their own ResolvedMethod
hierarchies, which are quite complicated and one of the few multiple
inheritance hierarchies in the compiler. But the compiler extensions
we use in the project (jitbuilder, compilertest) both basically do the
same things. Rather than have the compiler component define some APIs
with "UNIMPLEMENTED" bits that are typically implemented the same way,
this commit shifts all the code down into the compiler itself so that
common extensions do not have to define any special ResolvedMethod
classes (while leaving the opportunity open for other compiler extensions
to add/override with their own behaviour).

This commit moves the original compile/ResolvedMethod class, which
really creates TR_ResolvedMethod, into compile/TRResolvedMethod.[ch]pp.
It then moves the jitbuilder ResolvedMethod class (somewhat cleaned up)
into compile/OMRResolvedMethod to serve as the OMR::ResolvedMethod
base implementation. TR::ResolvedMethod now behaves more like an
extensible class in that compile/ResolvedMethod.hpp is its header
file, and it leverages OMRResolvedMethod to declare OMR::ResolvedMethod.
The slight strangeness is that OMR::ResolvedMethod inherits from
TR_ResolvedMethod and this class is still present for other projects,
like OpenJ9, that will not directly use OMR::ResolvedMethod but
build on TR_ResolvedMethod (sigh, can't fix everything at the same time).
This change also removes some vestigial artifacts of the original
refactoring of this class from the J9 implementation. For example,
JitBuilder::ResolvedMethodBase isn't even a useful class for
JitBuilder, let alone as part of the extensible ResolvedMethod
hierarchy. ResolvedMethodBase is eliminated as part of this change
after moving its implementation to other classes.

Some default implementations of in the OMR::Method class have also
been augmented to remove "UNIMPLEMENTED" and provide simple defaults
(for example, isConstructor() now returns false by default) so these
instance functions don't need to be implemented by compiler extensions.

One peculiarity that will be fixed in a subsequent commit is that
OMR::ResolvedMethod has some dependency on TR::IlType which is a
JitBuilder class. This will be removed in future by reducing these
types to TR::DataTypes.

With these changes, jitbuilder and compilertest no longer need to
have 3 files in the compile/ directory and so the compile/ directory
itself can be eliminated, simplifying the effort to create a new
compiler extension.

By removing JitBuilder and compilertest specific kinds of methods,
the set of Types for methods can be reduced to J9 and NonJ9. Since
the OpenJ9 project should use J9 specifically, we change the default
Type for a TR::Method from J9 to NonJ9.
