+++ compiler/compile/CMakeLists.txt
@@ -23,11 +23,12 @@ compiler_library(compile
 
 	${CMAKE_CURRENT_LIST_DIR}/OSRData.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRMethod.cpp
-	${CMAKE_CURRENT_LIST_DIR}/ResolvedMethod.cpp
+	${CMAKE_CURRENT_LIST_DIR}/OMRResolvedMethod.cpp
 	${CMAKE_CURRENT_LIST_DIR}/VirtualGuard.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRSymbolReferenceTable.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRAliasBuilder.cpp
 	${CMAKE_CURRENT_LIST_DIR}/OMRCompilation.cpp
 	${CMAKE_CURRENT_LIST_DIR}/TLSCompilationManager.cpp
+	${CMAKE_CURRENT_LIST_DIR}/TRResolvedMethod.cpp
 
 )
