+++ compiler/compile/OMRMethod.cpp
@@ -79,53 +79,27 @@ TR::ILOpCodes OMR::Method::returnOpCode()
     return TR::BadILOp;
 }
 
-uint16_t OMR::Method::classNameLength()
-{
-    TR_UNIMPLEMENTED();
-    return 0;
-}
+uint16_t OMR::Method::classNameLength() { return static_cast<uint16_t>(strlen(classNameChars())); }
 
-uint16_t OMR::Method::nameLength()
-{
-    TR_UNIMPLEMENTED();
-    return 0;
-}
+uint16_t OMR::Method::nameLength() { return static_cast<uint16_t>(strlen(nameChars())); }
 
-uint16_t OMR::Method::signatureLength()
-{
-    TR_UNIMPLEMENTED();
-    return 0;
-}
+uint16_t OMR::Method::signatureLength() { return static_cast<uint16_t>(strlen(signatureChars())); }
 
 char *OMR::Method::classNameChars()
 {
-    TR_UNIMPLEMENTED();
-    return 0;
+    return ""; // easy default
 }
 
 char *OMR::Method::nameChars()
 {
-    TR_UNIMPLEMENTED();
-    return 0;
+    return "Method"; // easy default
 }
 
-char *OMR::Method::signatureChars()
-{
-    TR_UNIMPLEMENTED();
-    return 0;
-}
+char *OMR::Method::signatureChars() { return "()V"; }
 
-bool OMR::Method::isConstructor()
-{
-    TR_UNIMPLEMENTED();
-    return false;
-}
+bool OMR::Method::isConstructor() { return false; }
 
-bool OMR::Method::isFinalInObject()
-{
-    TR_UNIMPLEMENTED();
-    return false;
-}
+bool OMR::Method::isFinalInObject() { return false; }
 
 const char *OMR::Method::signature(TR_Memory *, TR_AllocationKind)
 {
