+++ fvtest/compilertest/tests/TestDriver.cpp
@@ -40,7 +40,14 @@ TestDriver::RunTest()
 int32_t
 TestDriver::compileMethodBuilder(TR::MethodBuilder *m, uint8_t ** entry)
    {
-   TR::ResolvedMethod resolvedMethod(m);
+   TR::ResolvedMethod resolvedMethod((char *)m->getDefiningFile(),
+                                     (char *)m->getDefiningLine(),
+                                     (char *)m->GetMethodName(),
+                                     m->getNumParameters(),
+                                     m->getParameterTypes(),
+                                     m->getReturnType(),
+                                     0,
+                                     static_cast<TR::IlInjector *>(m));
    TR::IlGeneratorMethodDetails details(&resolvedMethod);
 
    int32_t rc=0;
